backend/requirements.txt
Flask==3.0.2
Werkzeug==3.0.2
Flask-Cors==4.0.1
flask-restful==0.3.9
python-dotenv==1.0.1
pymongo==4.7.2
requests==2.32.0
bcrypt==4.1.3
PyJWT==2.8.0